studymate-backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pymongo==4.6.0
google-auth==2.25.0
google-auth-oauthlib==1.2.0
python-dotenv==1.0.0
werkzeug==3.0.1
watchdog==3.0.0
spacy==3.7.2
transformers==4.35.0
torch==2.1.0
nltk==3.8.1
scikit-learn==1.3.2
sentence-transformers==2.2.2